refactor: move PaymentModal component to customer payment subdirectory
- Move PaymentModal.jsx from components/ to components/customer/payment/
- Update import statements in checkout-destination and checkout-package pages
- Maintain consistent organization pattern with other customer components
